Clean-up MediaPlaybackService
